refactor crap to fix circular imports
