+++ modules/sd_models.py
@@ -1,14 +1,20 @@
 import collections
 import os.path
 import sys
-from collections import namedtuple
 import torch
-from omegaconf import OmegaConf
 
+import modules.shared_steps.options as shared_opts
+import modules.devices as devices
+import modules.shared as shared
+
+from collections import namedtuple
+from omegaconf import OmegaConf
 from ldm.util import instantiate_from_config
 
-from modules import shared, modelloader, devices, script_callbacks
+from modules.command_options.options import cmd_opts, default_sd_model_file
+from modules.modelloader import load_models
 from modules.paths import models_path
+from modules.script_callbacks import model_loaded_callback
 from modules.sd_hijack_inpainting import do_inpainting_hijack, should_hijack_inpainting
 
 model_dir = "Stable-diffusion"
@@ -21,7 +27,7 @@ checkpoints_loaded = collections.OrderedDict()
 try:
     # this silences the annoying "Some weights of the model checkpoint were not used when initializing..." message at start.
 
-    from transformers import logging, CLIPModel
+    from transformers import logging
 
     logging.set_verbosity_error()
 except Exception:
@@ -41,13 +47,13 @@ def checkpoint_tiles():
 
 def list_models():
     checkpoints_list.clear()
-    model_list = modelloader.load_models(model_path=model_path, command_path=shared.cmd_opts.ckpt_dir, ext_filter=[".ckpt"])
+    model_list = load_models(model_path=model_path, command_path=cmd_opts.ckpt_dir, ext_filter=[".ckpt"])
 
     def modeltitle(path, shorthash):
         abspath = os.path.abspath(path)
 
-        if shared.cmd_opts.ckpt_dir is not None and abspath.startswith(shared.cmd_opts.ckpt_dir):
-            name = abspath.replace(shared.cmd_opts.ckpt_dir, '')
+        if cmd_opts.ckpt_dir is not None and abspath.startswith(cmd_opts.ckpt_dir):
+            name = abspath.replace(cmd_opts.ckpt_dir, '')
         elif abspath.startswith(model_path):
             name = abspath.replace(model_path, '')
         else:
@@ -60,13 +66,13 @@ def list_models():
 
         return f'{name} [{shorthash}]', shortname
 
-    cmd_ckpt = shared.cmd_opts.ckpt
+    cmd_ckpt = cmd_opts.ckpt
     if os.path.exists(cmd_ckpt):
         h = model_hash(cmd_ckpt)
         title, short_model_name = modeltitle(cmd_ckpt, h)
-        checkpoints_list[title] = CheckpointInfo(cmd_ckpt, title, h, short_model_name, shared.cmd_opts.config)
-        shared.opts.data['sd_model_checkpoint'] = title
-    elif cmd_ckpt is not None and cmd_ckpt != shared.default_sd_model_file:
+        checkpoints_list[title] = CheckpointInfo(cmd_ckpt, title, h, short_model_name, cmd_opts.config)
+        shared_opts.opts.data['sd_model_checkpoint'] = title
+    elif cmd_ckpt is not None and cmd_ckpt != default_sd_model_file:
         print(f"Checkpoint in --ckpt argument not found (Possible it was moved to {model_path}: {cmd_ckpt}", file=sys.stderr)
     for filename in model_list:
         h = model_hash(filename)
@@ -75,7 +81,7 @@ def list_models():
         basename, _ = os.path.splitext(filename)
         config = basename + ".yaml"
         if not os.path.exists(config):
-            config = shared.cmd_opts.config
+            config = cmd_opts.config
 
         checkpoints_list[title] = CheckpointInfo(filename, title, h, short_model_name, config)
 
@@ -101,18 +107,18 @@ def model_hash(filename):
 
 
 def select_checkpoint():
-    model_checkpoint = shared.opts.sd_model_checkpoint
+    model_checkpoint = shared_opts.opts.sd_model_checkpoint
     checkpoint_info = checkpoints_list.get(model_checkpoint, None)
     if checkpoint_info is not None:
         return checkpoint_info
 
     if len(checkpoints_list) == 0:
         print(f"No checkpoints found. When searching for checkpoints, looked at:", file=sys.stderr)
-        if shared.cmd_opts.ckpt is not None:
-            print(f" - file {os.path.abspath(shared.cmd_opts.ckpt)}", file=sys.stderr)
+        if cmd_opts.ckpt is not None:
+            print(f" - file {os.path.abspath(cmd_opts.ckpt)}", file=sys.stderr)
         print(f" - directory {model_path}", file=sys.stderr)
-        if shared.cmd_opts.ckpt_dir is not None:
-            print(f" - directory {os.path.abspath(shared.cmd_opts.ckpt_dir)}", file=sys.stderr)
+        if cmd_opts.ckpt_dir is not None:
+            print(f" - directory {os.path.abspath(cmd_opts.ckpt_dir)}", file=sys.stderr)
         print(f"Can't run without a checkpoint. Find and place a .ckpt file into any of those locations. The program will exit.", file=sys.stderr)
         exit(1)
 
@@ -172,19 +178,19 @@ def load_model_weights(model, checkpoint_info):
         sd = get_state_dict_from_checkpoint(pl_sd)
         missing, extra = model.load_state_dict(sd, strict=False)
 
-        if shared.cmd_opts.opt_channelslast:
+        if cmd_opts.opt_channelslast:
             model.to(memory_format=torch.channels_last)
 
-        if not shared.cmd_opts.no_half:
+        if not cmd_opts.no_half:
             model.half()
 
-        devices.dtype = torch.float32 if shared.cmd_opts.no_half else torch.float16
-        devices.dtype_vae = torch.float32 if shared.cmd_opts.no_half or shared.cmd_opts.no_half_vae else torch.float16
+        devices.dtype = torch.float32 if cmd_opts.no_half else torch.float16
+        devices.dtype_vae = torch.float32 if cmd_opts.no_half or cmd_opts.no_half_vae else torch.float16
 
         vae_file = os.path.splitext(checkpoint_file)[0] + ".vae.pt"
 
-        if not os.path.exists(vae_file) and shared.cmd_opts.vae_path is not None:
-            vae_file = shared.cmd_opts.vae_path
+        if not os.path.exists(vae_file) and cmd_opts.vae_path is not None:
+            vae_file = cmd_opts.vae_path
 
         if os.path.exists(vae_file):
             print(f"Loading VAE weights from: {vae_file}")
@@ -195,7 +201,7 @@ def load_model_weights(model, checkpoint_info):
         model.first_stage_model.to(devices.dtype_vae)
 
         checkpoints_loaded[checkpoint_info] = model.state_dict().copy()
-        while len(checkpoints_loaded) > shared.opts.sd_checkpoint_cache:
+        while len(checkpoints_loaded) > shared_opts.opts.sd_checkpoint_cache:
             checkpoints_loaded.popitem(last=False)  # LRU
     else:
         print(f"Loading weights [{sd_model_hash}] from cache")
@@ -211,7 +217,7 @@ def load_model(checkpoint_info=None):
     from modules import lowvram, sd_hijack
     checkpoint_info = checkpoint_info or select_checkpoint()
 
-    if checkpoint_info.config != shared.cmd_opts.config:
+    if checkpoint_info.config != cmd_opts.config:
         print(f"Loading config from: {checkpoint_info.config}")
 
     sd_config = OmegaConf.load(checkpoint_info.config)
@@ -230,8 +236,8 @@ def load_model(checkpoint_info=None):
     sd_model = instantiate_from_config(sd_config.model)
     load_model_weights(sd_model, checkpoint_info)
 
-    if shared.cmd_opts.lowvram or shared.cmd_opts.medvram:
-        lowvram.setup_for_low_vram(sd_model, shared.cmd_opts.medvram)
+    if cmd_opts.lowvram or cmd_opts.medvram:
+        lowvram.setup_for_low_vram(sd_model, cmd_opts.medvram)
     else:
         sd_model.to(shared.device)
 
@@ -240,7 +246,7 @@ def load_model(checkpoint_info=None):
     sd_model.eval()
     shared.sd_model = sd_model
 
-    script_callbacks.model_loaded_callback(sd_model)
+    model_loaded_callback(sd_model)
 
     print(f"Model loaded.")
     return sd_model
@@ -258,7 +264,7 @@ def reload_model_weights(sd_model, info=None):
         load_model(checkpoint_info)
         return shared.sd_model
 
-    if shared.cmd_opts.lowvram or shared.cmd_opts.medvram:
+    if cmd_opts.lowvram or cmd_opts.medvram:
         lowvram.send_everything_to_cpu()
     else:
         sd_model.to(devices.cpu)
@@ -268,9 +274,9 @@ def reload_model_weights(sd_model, info=None):
     load_model_weights(sd_model, checkpoint_info)
 
     sd_hijack.model_hijack.hijack(sd_model)
-    script_callbacks.model_loaded_callback(sd_model)
+    model_loaded_callback(sd_model)
 
-    if not shared.cmd_opts.lowvram and not shared.cmd_opts.medvram:
+    if not cmd_opts.lowvram and not cmd_opts.medvram:
         sd_model.to(devices.device)
 
     print(f"Weights loaded.")
