+++ modules/shared_steps/options.py
@@ -1,0 +1,94 @@
+import json
+import sys
+
+import modules.shared_steps.templates as shared_templates
+
+class Options:
+    data = None
+    data_labels = shared_templates.options_templates
+    typemap = {int: float}
+
+    def __init__(self):
+        self.data = {k: v.default for k, v in self.data_labels.items()}
+
+    def __setattr__(self, key, value):
+        if self.data is not None:
+            if key in self.data:
+                self.data[key] = value
+
+        return super(Options, self).__setattr__(key, value)
+
+    def __getattr__(self, item):
+        if self.data is not None:
+            if item in self.data:
+                return self.data[item]
+
+        if item in self.data_labels:
+            return self.data_labels[item].default
+
+        return super(Options, self).__getattribute__(item)
+
+    def save(self, filename):
+        with open(filename, "w", encoding="utf8") as file:
+            json.dump(self.data, file, indent=4)
+
+    def same_type(self, x, y):
+        if x is None or y is None:
+            return True
+
+        type_x = self.typemap.get(type(x), type(x))
+        type_y = self.typemap.get(type(y), type(y))
+
+        return type_x == type_y
+
+    def load(self, filename):
+        with open(filename, "r", encoding="utf8") as file:
+            self.data = json.load(file)
+
+        bad_settings = 0
+        for k, v in self.data.items():
+            info = self.data_labels.get(k, None)
+            if info is not None and not self.same_type(info.default, v):
+                print(
+                    f"Warning: bad setting value: {k}: {v} ({type(v).__name__}; expected {type(info.default).__name__})",
+                    file=sys.stderr,
+                )
+                bad_settings += 1
+
+        if bad_settings > 0:
+            print(
+                f"The program is likely to not work with bad settings.\nSettings file: {filename}\nEither fix the file, or delete it and restart.",
+                file=sys.stderr,
+            )
+
+    def onchange(self, key, func):
+        item = self.data_labels.get(key)
+        item.onchange = func
+
+        func()
+
+    def dumpjson(self):
+        d = {
+            k: self.data.get(k, self.data_labels.get(k).default)
+            for k in self.data_labels.keys()
+        }
+        return json.dumps(d)
+
+    def add_option(self, key, info):
+        self.data_labels[key] = info
+
+    def reorder(self):
+        """reorder settings so that all items related to section always go together"""
+
+        section_ids = {}
+        settings_items = self.data_labels.items()
+        for k, item in settings_items:
+            if item.section not in section_ids:
+                section_ids[item.section] = len(section_ids)
+
+        self.data_labels = {
+            k: v
+            for k, v in sorted(settings_items, key=lambda x: section_ids[x[1].section])
+        }
+
+opts = Options()
