+++ modules/txt2img.py
@@ -1,15 +1,15 @@
 import modules.scripts
-from modules.processing import StableDiffusionProcessing, Processed, StableDiffusionProcessingTxt2Img, \
-    StableDiffusionProcessingImg2Img, process_images
-from modules.shared import opts, cmd_opts
-import modules.shared as shared
-import modules.processing as processing
+
+from modules.command_options.options import cmd_opts
+from modules.shared_steps.options import opts
+from modules.processing import StableDiffusionProcessingTxt2Img, process_images
+from modules.shared import sd_model, progress_print_out, total_tqdm
 from modules.ui import plaintext_to_html
 
 
 def txt2img(prompt: str, negative_prompt: str, prompt_style: str, prompt_style2: str, steps: int, sampler_index: int, restore_faces: bool, tiling: bool, n_iter: int, batch_size: int, cfg_scale: float, seed: int, subseed: int, subseed_strength: float, seed_resize_from_h: int, seed_resize_from_w: int, seed_enable_extras: bool, height: int, width: int, enable_hr: bool, denoising_strength: float, firstphase_width: int, firstphase_height: int, *args):
     p = StableDiffusionProcessingTxt2Img(
-        sd_model=shared.sd_model,
+        sd_model=sd_model,
         outpath_samples=opts.outdir_samples or opts.outdir_txt2img_samples,
         outpath_grids=opts.outdir_grids or opts.outdir_txt2img_grids,
         prompt=prompt,
@@ -40,14 +40,14 @@ def txt2img(prompt: str, negative_prompt: str, prompt_style: str, prompt_style2:
     p.script_args = args
 
     if cmd_opts.enable_console_prompts:
-        print(f"\ntxt2img: {prompt}", file=shared.progress_print_out)
+        print(f"\ntxt2img: {prompt}", file=progress_print_out)
 
     processed = modules.scripts.scripts_txt2img.run(p, *args)
 
     if processed is None:
         processed = process_images(p)
 
-    shared.total_tqdm.clear()
+    total_tqdm.clear()
 
     generation_info_js = processed.js()
     if opts.samples_log_stdout:
