+++ run_webui_mac.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 
 # Run the web ui
-python webui.py --precision full --no-half --use-cpu Interrogate GFPGAN CodeFormer BSRGAN ESRGAN SCUNet
+python3 webui.py --precision full --no-half --use-cpu Interrogate GFPGAN CodeFormer ESRGAN SCUNet
