+++ scripts/prompts_from_file.py
@@ -1,6 +1,4 @@
 import copy
-import math
-import os
 import random
 import sys
 import traceback
@@ -10,8 +8,7 @@ import modules.scripts as scripts
 import gradio as gr
 
 from modules.processing import Processed, process_images
-from PIL import Image
-from modules.shared import opts, cmd_opts, state
+from modules.shared import state
 
 
 def process_string_tag(tag):
